+++ lib/yabeda/rspec/increment_yabeda_counter.rb
@@ -15,6 +15,7 @@ module Yabeda
     class IncrementYabedaCounter < BaseMatcher
       def by(increment)
         @expected_increment = increment
+        @expectations = { tags => increment } if tags
         self
       end
 
@@ -32,8 +33,12 @@ module Yabeda
 
         increments = filter_matching_changes(Yabeda::TestAdapter.instance.counters.fetch(metric))
 
-        increments.values.any? do |actual_increment|
-          expected_increment.nil? || values_match?(expected_increment, actual_increment)
+        return false if increments.empty?
+
+        increments.values.all? do |expected_increment, actual_increment|
+          next !actual_increment.nil? if expected_increment.nil?
+
+          values_match?(expected_increment, actual_increment)
         end
       end
 
@@ -49,13 +54,16 @@ module Yabeda
 
         increments = filter_matching_changes(Yabeda::TestAdapter.instance.counters.fetch(metric))
 
-        increments.none?
+        increments.none? { |_tags, (_expected, actual)| !actual.nil? }
       end
 
       def failure_message
         "expected #{expected_formatted} " \
           "to be incremented #{"by #{description_of(expected_increment)} " unless expected_increment.nil?}" \
           "#{"with tags #{::RSpec::Support::ObjectFormatter.format(tags)} " if tags}" \
+          "#{if !tags && expectations
+               "with following expectations: #{::RSpec::Support::ObjectFormatter.format(expectations)} "
+             end}" \
           "but #{actual_increments_message}"
       end
 
@@ -63,6 +71,9 @@ module Yabeda
         "expected #{expected_formatted} " \
           "not to be incremented " \
           "#{"with tags #{::RSpec::Support::ObjectFormatter.format(tags)} " if tags}" \
+          "#{if !tags && expectations
+               "with following expectations: #{::RSpec::Support::ObjectFormatter.format(expectations)} "
+             end}" \
           "but #{actual_increments_message}"
       end